Update patch set 2
Patch Set 2:

(2 comments)

Hi Jake,

It's nice to be on-net, but the amount of actual traffic we're talking about here won't be much. When I looked at all the official mirrors, it's only Optus that off-net, so I'm not concerned.

I think the tech burden of handling our own mirror redirect means it's not worth it.

My rationale is that it's better to go with default mirrors that should *work* rather than break, but guarantee we're on-net.

Patch-set: 2
Reviewer: Gerrit User 1000077 <1000077@f182a8af-62ca-4ceb-ab81-14f8fec9c6c5>
Label: Verified=0