fix fail2ban script when VERSION_ID contains '.'
This is already fixed in one spot in this script, this patch fixes
it in other places. If VERSION_ID=7.2, then the script silently
fails. This occurs on most RHEL derivatives, but not on CentOS,
however there is a CentOS bug report open here:

https://bugs.centos.org/view.php?id=8359

So it may get fixed in the future on CentOS.

Change-Id: Iaa57f741d3c26f8779c5c572221f422b6df4d2a1