disable reverse DNS lookups upon login for speedup
When logging into a server configured with `UseDNS yes`, the server
performs a reverse DNS lookup of the client's IP. If the client has
no reverse DNS, this can cause a delay logging in of up to 30 seconds.

This is only required for redhat-based distros, as suse, debian and
ubuntu all have `UseDNS no` as the default.

Change-Id: I155c37031d2832b3ca72c3adf5e612a9df50262f