Update patch set 1
Patch Set 1: Code-Review-1

I am conflicted about this... the long timeout is usually caused by a broken DNS and not NXDOMAIN. Working around this at ssh login does not fix the problem; however a broken DNS might cause services dependent on it to behave weirdly, which can be quite hard to debug.

I would rather the user have a visible indication at login if his DNS is broken, so that he is more inclined to fix it.

Do you have an example where DNS is currently broken on our cloud?

Patch-set: 1