Update patch set 1
Patch Set 1:

It's not the DNS on the cloud, it's a DNS lookup in the VM of the IP that the client is connecting from.

The new default in upstream openssh is to use "UseDNS no" because not all IP addresses (clients) have reverse DNS entries. Red Hat is just very behind on their openssh version.

Patch-set: 1
Reviewer: Gerrit User 1000147 <1000147@f182a8af-62ca-4ceb-ab81-14f8fec9c6c5>
Label: Verified=0