Update patch set 1
Patch Set 1:

Yes I understand. This can be caused by two scenerios
1) DNS servers in the image are broken
2) DNS servers for the client's PTR record are broken
If it's (1) we have to fix it.
If it's (2) the client should fix it, or know that it's broken so that he can disable it explicitly. This also help him to understand what is wrong in case he uses other remote tools that might have the same behaviour, e.g. ftp, x2go..

UseDNS no is a good idea but I am also wary of deviating too much from default OS behaviour out of the box.

More people should vote on this.

Patch-set: 1