Update patch set 1
Patch Set 1:

Upstream openssh changed to `UseDNS no` in openssh 6.8.
Upstream Red Hat will include `UseDNS no` in RHEL8 (and CentOS 8 will also get it)
Fedora already has `UseDNS no`, as do Ubuntu and Debian and openSUSE.

It might be difficult for users to get their ISP or Telco to fix the DNS for the IP address from which they are connecting, which is why I think the default has changed in all distros.

Patch-set: 1