Update patch set 1
Patch Set 1:

So I've had a look, and I don't think it's worth while doing anything with heat-cfntools. It's a tiny little package with no service script or anything, so not much benefit in rearchitecting just to remove it. It's harmless.

So I've kept cloud as is, as the only changes we make are: lock metadata to OpenStack and provide more generous timeouts for metadata. To install cloud-init some other way would require another role just for it, which isn't worth it.

ntp and repos has now been removed.

Patch-set: 1
Label: Verified=0