Remove dedicated Murano images
We can now drop dedicated murano images as we can just start the
murano agent when needed, after a change to the murano provisioning
script recently committed.

This commit also sets and image build dir, which is used by Ansible
for dumping facts that later get added to the final image
dynamically.

Change-Id: I700af658cef310aaec782e0c1922950e50953d25