Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> Can we use a bind mount instead of mounting it in different locations? It might help with the clean up logic.
> 
> (-1 for visibility)

Jakey: that would be up to the user, not us to do.  If they understood what a bind mount.  But I am thinking that we ought to support multiple mounts anyway, and if we do that then multiple mountpoints and -m are essential.

Patch-set: 2
Reviewer: Gerrit User 1000157 <1000157@f182a8af-62ca-4ceb-ab81-14f8fec9c6c5>