Add support for automatic security upgrades
This commit adds support for all supported distros using:
  - Unattended upgrades (Debian and Ubuntu)
  - DNF Automatic (Fedora and CentOS 8+)
  - YUM cron (CentOS 7)

Each are configured to automatically upgrade packages marked as
"security" updates, except CentOS where it doesn't know which
are security, so CentOS will update all packages, except in the
case of CentOS Stream, where updates are disables.

An image property will be added to indicate this feature is enabled
and whether it's set up for security or all updates.

Change-Id: I0b4d500de63719c36c1ec9f3674c568e7fd79c2a