Add glibc-langpack-en for CentOS Stream 9 build
As mentioned in the ticket:
  https://support.ehelp.edu.au/a/tickets/201546
CentOS Stream 9 shows some locale errors from some operations.

We actually encountered this (and solved it) for Alma Linux 9
too in commmit 653e74c13d7dd1fc2222fe53b07ad7895e30c86a

This might potentially be useful for Rocky Linux too, but this
is not confirmed yet.

Change-Id: I0642bb344682695da37ac67487ac31b8d21cbc50