Split out Firefox into its own role
For Ubuntu, we need to remove the snap version of Firefox and
install the Mozilla PPA deb version instead.

With these extra steps, and that Firefox is needed for a few
different scenarios, it made sense to split it into its own module.

The firefox module is now set as a dependency for desktop-common,
jenkins (selenium), xfce4-desktop and neurodesk (added shortly)

Change-Id: I69ef88a2036fcbc9a75b1883679563c504142f28